add new file

+++ src/main/java/jdbc/JDBCTest.java
@@ -17,13 +17,17 @@ public class JDBCTest {
         String url = "jdbc:mysql://localhost:3306/jdbcdemo";
         String user = "root";
         String password = "2016";
+        // Register the Driver class
         Class.forName("com.mysql.cj.jdbc.Driver");
+        // Create connection
         try (Connection conn = DriverManager.getConnection(url, user, password)) {
+            // Create statement
             try (Statement stmt = conn.createStatement()) {
+                // Execute queries
                 try (ResultSet rs = stmt.executeQuery("select id, grade, name, gender from students where gender=1")) {
                     while (rs.next()) {
                         long id = rs.getLong(1);
-                        long grade = rs.getLong(2);
+                        int grade = rs.getInt(2);
                         String name = rs.getString(3);
                         int gender = rs.getInt(4);
                         System.out.println("id = " + id + ", grade = " + grade + ", name = " + name + ", gender = " + gender);
